+++ pypgx/cli/import_variants.py
@@ -6,7 +6,12 @@ import fuc
 import pysam
 
 description = f"""
-Import variant data for the target gene.
+Import variant (SNV/indel) data for the target gene.
+
+The command will first slice input VCF for the target gene and then assess
+whether every genotype call in the sliced VCF is haplotype phased. It will
+return an archive file with the semantic type VcfFrame[Consolidated] if the
+VCF is fully phased or otherwise VcfFrame[Imported].
 """
 
 def create_parser(subparsers):
@@ -14,7 +19,7 @@ def create_parser(subparsers):
         subparsers,
         fuc.api.common._script_name(),
         description=description,
-        help='Import variant data for the target gene.',
+        help='Import variant (SNV/indel) data for the target gene',
     )
     parser.add_argument(
         'gene',
@@ -27,7 +32,8 @@ def create_parser(subparsers):
     parser.add_argument(
         'imported_variants',
         metavar='imported-variants',
-        help='Archive file with the semantic type VcfFrame[Imported].'
+        help='Archive file with the semantic type VcfFrame[Imported] \n'
+             'or VcfFrame[Consolidated].'
     )
     parser.add_argument(
         '--assembly',